Return `impl Future` instead of `Pin<Box<dyn Future>>`
Returning a boxed trait object requires specifying whether the returned
future is `Send`/`Sync`, which means we can't support both `Send` and
non-`Send` async HTTP clients. The `reqwest` client returns a `Send`
future when using versions of `hyper` >= 0.14.14 (see
https://github.com/hyperium/hyper/pull/2653), except in WASM builds,
for which the future is non-`Send`.

With this change, each `request_async()` method now returns
`impl Future`, which may or may not be `Send` depending on the inputs.
This provides the flexibility to support all the necessary HTTP
clients, while providing the convenience of `Send` futures for the
built-in `reqwest` client in non-WASM builds.

BREAKING CHANGES:
 - Removed `TokenRequestFuture` and `DeviceAuthorizationRequestFuture`
   type aliases.
 - Added `Future` associated type to `AsyncHttpClient` trait and changed
   `call()` method to return `Self::Future` (which may or may not be
   `Send`/`Sync`).
 - Changed each `request_async()` method to return `impl Future` instead
   of `Pin<Box<dyn Future>>`.

Fixes #255 and #260.